pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.6",
    "pillow>=11.3.0",
    "pydantic-ai>=1.0.1",
    "python-dotenv>=1.1.1",
]
